[FIX] code comment & [RENAME]
